Update local K8s configuration to be runnable on any cluster.
This is prefactoring to support running K8s tests against clusters running in cloud environments.

Note that as a result, the correctness test target no longer manages creation/deletion of the cluster. Instead, a cluster as well as a container registry that is accessible from that cluster are expected to already exist.
