+++ src/main/k8s/dev/resource_setup_gke.cue
@@ -18,14 +18,9 @@ _secret_name: string @tag("secret_name")
 
 objectSets: [resourceSetup]
 
-_imageConfig: #ImageConfig & {
-	repoSuffix: "loadtest/resource-setup"
-}
-
 resourceSetup: #ResourceSetup & {
 	_edp_display_names: ["edp1", "edp2", "edp3", "edp4", "edp5", "edp6"]
 	_duchy_ids: ["aggregator", "worker1", "worker2"]
-	_job_image:                  _imageConfig.image
 	_resource_setup_secret_name: _secret_name
 	_dependencies: ["gcp-kingdom-data-server"]
 }
