+++ src/main/k8s/edp_simulator.cue
@@ -28,8 +28,9 @@ package k8s
 
 	let DisplayName = _edpConfig.displayName
 
-	_edp_simulator_image:         string
-	_simulator_image_pull_policy: string
+	_imageConfig: #ImageConfig & {
+		repoSuffix: string | *"loadtest/edp-simulator"
+	}
 	_blob_storage_flags: [...string]
 
 	_additional_args: [...string]
@@ -39,22 +40,21 @@ package k8s
 		_secretName: _edp_secret_name
 		_system:     "simulator"
 		_container: {
-			image:           _edp_simulator_image
-			imagePullPolicy: _simulator_image_pull_policy
-			args:            [
-						"--tls-cert-file=/var/run/secrets/files/\(DisplayName)_tls.pem",
-						"--tls-key-file=/var/run/secrets/files/\(DisplayName)_tls.key",
-						"--cert-collection-file=/var/run/secrets/files/all_root_certs.pem",
-						"--data-provider-resource-name=\(_edpConfig.resourceName)",
-						"--data-provider-display-name=\(DisplayName)",
-						"--data-provider-encryption-private-keyset=/var/run/secrets/files/\(DisplayName)_enc_private.tink",
-						"--data-provider-consent-signaling-private-key-der-file=/var/run/secrets/files/\(DisplayName)_cs_private.der",
-						"--data-provider-consent-signaling-certificate-der-file=/var/run/secrets/files/\(DisplayName)_cs_cert.der",
-						"--mc-resource-name=\(_mc_resource_name)",
-						"--kingdom-public-api-target=\(_kingdom_public_api_target)",
-						"--kingdom-public-api-cert-host=localhost",
-						"--requisition-fulfillment-service-target=\(_duchy_public_api_target)",
-						"--requisition-fulfillment-service-cert-host=localhost",
+			image: _imageConfig.image
+			args:  [
+				"--tls-cert-file=/var/run/secrets/files/\(DisplayName)_tls.pem",
+				"--tls-key-file=/var/run/secrets/files/\(DisplayName)_tls.key",
+				"--cert-collection-file=/var/run/secrets/files/all_root_certs.pem",
+				"--data-provider-resource-name=\(_edpConfig.resourceName)",
+				"--data-provider-display-name=\(DisplayName)",
+				"--data-provider-encryption-private-keyset=/var/run/secrets/files/\(DisplayName)_enc_private.tink",
+				"--data-provider-consent-signaling-private-key-der-file=/var/run/secrets/files/\(DisplayName)_cs_private.der",
+				"--data-provider-consent-signaling-certificate-der-file=/var/run/secrets/files/\(DisplayName)_cs_cert.der",
+				"--mc-resource-name=\(_mc_resource_name)",
+				"--kingdom-public-api-target=\(_kingdom_public_api_target)",
+				"--kingdom-public-api-cert-host=localhost",
+				"--requisition-fulfillment-service-target=\(_duchy_public_api_target)",
+				"--requisition-fulfillment-service-cert-host=localhost",
 			] + _blob_storage_flags + _additional_args
 		}
 	}
