+++ src/main/k8s/local/db_creds_kustomization.yaml
@@ -1,4 +1,4 @@
-# Copyright 2022 The Cross-Media Measurement Authors
+# Copyright 2023 The Cross-Media Measurement Authors
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -12,7 +12,8 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-configMapGenerator:
-- name: config-files
-  files:
-  - authority_key_identifier_to_principal_map.textproto
+secretGenerator:
+- name: db-creds
+  literals:
+  - username=db-user
+  - password=hunter2
