+++ src/main/k8s/local/empty_akid_mapping_config.textproto
