+++ src/main/k8s/local/empty_encryption_key_pair_config.textproto
@@ -1,0 +1,2 @@
+# proto-file: wfa/measurement/config/reporting/encryption_key_pair_config.proto
+# proto-message: EncryptionKeyPairConfig
