+++ src/test/kotlin/org/wfanet/measurement/integration/k8s/BUILD.bazel
@@ -4,47 +4,64 @@ load("@rules_java//java:defs.bzl", "java_test")
 package(default_testonly = True)
 
 kt_jvm_library(
-    name = "correctness_test",
-    srcs = ["CorrectnessTest.kt"],
+    name = "abstract_correctness_test",
+    srcs = ["AbstractCorrectnessTest.kt"],
     data = [
-        "//src/main/k8s/local/testing:duchies_and_edps_config",
-        "//src/main/k8s/local/testing:emulators_config",
-        "//src/main/k8s/local/testing:kingdom_config",
         "//src/main/k8s/testing/secretfiles:mc_trusted_certs.pem",
         "//src/main/k8s/testing/secretfiles:secret_files",
     ],
     deps = [
+        "//src/main/kotlin/org/wfanet/measurement/integration/common:configs",
+        "//src/main/kotlin/org/wfanet/measurement/loadtest/frontend:frontend_simulator",
+        "@wfa_common_jvm//imports/java/com/google/common/truth",
+        "@wfa_common_jvm//imports/java/org/junit",
+        "@wfa_common_jvm//imports/kotlin/kotlinx/coroutines:core",
+        "@wfa_common_jvm//src/main/kotlin/org/wfanet/measurement/common",
+    ],
+)
+
+kt_jvm_library(
+    name = "empty_cluster_correctness_test",
+    srcs = ["EmptyClusterCorrectnessTest.kt"],
+    deps = [
+        ":abstract_correctness_test",
         "//imports/java/io/kubernetes/client",
         "//imports/java/io/kubernetes/client/extended",
-        "//imports/java/okhttp3",
         "//src/main/kotlin/org/wfanet/measurement/common/k8s",
-        "//src/main/kotlin/org/wfanet/measurement/common/k8s/testing:kind_cluster",
         "//src/main/kotlin/org/wfanet/measurement/common/k8s/testing:port_forwarder",
+        "//src/main/kotlin/org/wfanet/measurement/common/k8s/testing:processes",
         "//src/main/kotlin/org/wfanet/measurement/integration/common:configs",
         "//src/main/kotlin/org/wfanet/measurement/loadtest/config:event_filters",
         "//src/main/kotlin/org/wfanet/measurement/loadtest/frontend:frontend_simulator",
         "//src/main/kotlin/org/wfanet/measurement/loadtest/resourcesetup:resource_setup",
-        "@wfa_common_jvm//imports/java/com/google/common/truth",
         "@wfa_common_jvm//imports/java/org/jetbrains/annotations",
         "@wfa_common_jvm//imports/java/org/junit",
+        "@wfa_common_jvm//imports/kotlin/kotlinx/coroutines:core",
         "@wfa_common_jvm//imports/kotlin/kotlinx/coroutines:jdk8",
-        "@wfa_common_jvm//src/main/kotlin/org/wfanet/measurement/common",
+        "@wfa_common_jvm//src/main/kotlin/org/wfanet/measurement/common/crypto:security_provider",
+        "@wfa_common_jvm//src/main/kotlin/org/wfanet/measurement/common/grpc",
+        "@wfa_common_jvm//src/main/kotlin/org/wfanet/measurement/common/testing",
         "@wfa_common_jvm//src/main/kotlin/org/wfanet/measurement/storage/forwarded",
     ],
 )
 
 java_test(
-    name = "CorrectnessTest",
+    name = "EmptyClusterCorrectnessTest",
     size = "enormous",
     timeout = "long",
     data = [
-        "//src/main/k8s/local/testing:cmms_image_archives",
+        "//src/main/docker:push_all_local_images",
+        "//src/main/k8s/local:kingdom_setup.tar",
+        "//src/main/k8s/local/testing:cmms.tar",
     ],
-    env_inherit = ["DOCKER_HOST"],
+    env_inherit = [
+        "DOCKER_HOST",
+        "KUBECONFIG",
+    ],
     tags = [
         "cpu:8",
         "manual",
     ],
-    test_class = "org.wfanet.measurement.integration.k8s.CorrectnessTest",
-    runtime_deps = [":correctness_test"],
+    test_class = "org.wfanet.measurement.integration.k8s.EmptyClusterCorrectnessTest",
+    runtime_deps = [":empty_cluster_correctness_test"],
 )
